fix: Fix error handling and communication to user in book service
Book service was not providing adequate user communication on the various kinds of possible errors that could occur while fetching and parsing the book information from the remote server. Also, there was a bug that it did not exit even when there was an IO error and proceeded with the parsing of the json string. This caused a null pointer exception when there was an IO error.